requirements.txt
PySide2
PyMuPDF
Pillow